refactor: reorganize domain packages and move static-factory-method to infra module for separation of concerns
