+++ domain/src/main/java/hanghae/domain/types/movie/AgeRating.java
@@ -1,4 +1,4 @@
-package hanghae.domain.common.movie;
+package hanghae.domain.types.movie;
 
 public enum AgeRating {
     ALL, AGE_12, AGE_15, AGE_19;
