+++ domain/src/main/java/hanghae/domain/types/movie/Genre.java
@@ -1,4 +1,4 @@
-package hanghae.domain.common.movie;
+package hanghae.domain.types.movie;
 
 public enum Genre {
     ACTION, SF, ROMANCE, COMEDY, HORROR, DRAMA, ANIMATION
