+++ domain/src/main/java/hanghae/domain/types/movie/MovieId.java
@@ -1,4 +1,4 @@
-package hanghae.domain.common.movie;
+package hanghae.domain.types.movie;
 
 import java.util.Objects;
