+++ domain/src/main/java/hanghae/domain/types/movie/ReleaseDate.java
@@ -1,4 +1,4 @@
-package hanghae.domain.common.movie;
+package hanghae.domain.types.movie;
 
 
 import lombok.Getter;
